Enhance delete contact feature
The list of participants for a meeting is not updated even when the
corresponding contact is deleted.

This is problematic as the user needs to manually edit the meeting after
deleting a contact to update the participant list.

Let's enhance the delete contact feature so that deleting a contact
automatically updates the participant list of affected meetings.