Enhance update contact feature
The list of participants for a meeting is not updated even when the
corresponding contact is edited, so the information displayed in the
meeting's participant list may not always be accurate.

Let's enhance the update contact feature so that editing a contact
automatically updates the participant list of affected meetings.